# services/api/openapi.proposal.yaml
...
openapi: 3.1.0
info:
  title: TuneDeck API
  version: 0.2.0-proposal
servers:
  - url: /
paths:
...
  /v1/me:
    get:
      summary: The signed-in account (the account code customers read out to support)
      description: No OIDC subject and no roles. The email is the access token's email claim, stored at sign-in.
      responses:
        '200':
          description: Account
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MeAccount' } } }
        '401': { description: AUTH_REQUIRED, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ACCOUNT_DELETING, AUTH_FORBIDDEN (account disabled) or DEVICE_REVOKED (the token''s Keycloak session belongs to a signed-out device)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'API_RATE_LIMITED, 120 reads a minute per account; details.retryAfterSeconds', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { $ref: '#/components/responses/Error' }
    delete:
      summary: Start account deletion (Doc 17 path)
      description: >-
        Requires the token's auth_time to be within the last 300 seconds and an Idempotency-Key. From this call the
        account is locked (every later call is 403 ACCOUNT_DELETING), every device is signed out, and a purge is queued
        (Keycloak user first, then local data). Poll GET /v1/account-deletions/{ticket}. Audited account.delete_requested.
        A retry with the same Idempotency-Key after a success gets 403 ACCOUNT_DELETING, not the ticket again
        (the sign-in check refuses the locked account before the replay).
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: 'Required on this path (428 IDEMPOTENCY_KEY_REQUIRED without it). 8-128 characters of A-Z a-z 0-9 _ . : -, e.g. a UUID.'
          schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' }
      responses:
        '202':
          description: Deletion started. Keep the ticket; it is shown only once.
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MeDeletionTicket' } } }
        '400': { description: 'VALIDATION_FAILED, details.field = Idempotency-Key (malformed key)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED, or REAUTH_REQUIRED with details.maxAgeSeconds = 300', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ACCOUNT_DELETING (already requested), AUTH_FORBIDDEN or DEVICE_REVOKED', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 'IDEMPOTENCY_KEY_REUSED (same key, different request) or IDEMPOTENCY_IN_PROGRESS (Retry-After: 1)', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '428': { description: 'IDEMPOTENCY_KEY_REQUIRED, details.header = Idempotency-Key', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'API_RATE_LIMITED, 30 writes a minute per account; details.retryAfterSeconds', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { $ref: '#/components/responses/Error' }
  /v1/me/account:
    delete:
      summary: Start account deletion (older path, same as DELETE /v1/me)
      description: >-
        Same behaviour as DELETE /v1/me, except that Idempotency-Key is optional here (kept for app builds made before
        the key existed). Requires auth_time within the last 300 seconds. As there, a retry after a success gets
        403 ACCOUNT_DELETING rather than a replay of the ticket.
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Deletion started
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MeDeletionTicket' } } }
        '400': { description: 'VALIDATION_FAILED, details.field = Idempotency-Key (malformed key)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED, or REAUTH_REQUIRED with details.maxAgeSeconds = 300', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ACCOUNT_DELETING (already requested), AUTH_FORBIDDEN or DEVICE_REVOKED', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 'IDEMPOTENCY_KEY_REUSED or IDEMPOTENCY_IN_PROGRESS (Retry-After: 1)', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'API_RATE_LIMITED, 30 writes a minute per account; details.retryAfterSeconds', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { $ref: '#/components/responses/Error' }
  /v1/account-deletions/{ticket}:
    get:
      summary: Public progress of an account deletion (no sign-in)
      description: Rate limited to 60 a minute per client address. Unknown, malformed and expired tickets are all 404. Tickets are kept 35 days after completion.
      security: []
      parameters:
        - { name: ticket, in: path, required: true, description: The ticket from DELETE /v1/me, schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' } }
      responses:
        '200':
          description: Progress
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MeDeletionStatus' } } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { description: 'API_RATE_LIMITED, 60 a minute per client address; details.retryAfterSeconds', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { $ref: '#/components/responses/Error' }
...
  /v1/me/export:
    get:
      summary: The account's data built on the spot (legacy; new clients use POST /v1/me/exports)
      description: Needs a sign-in from the last 5 minutes (401 REAUTH_REQUIRED otherwise), since it returns all of the account's data. Audited account.export on every call. At most 10,000 diagnostic events (diagnosticsTruncated says when more exist).
      responses:
        '200':
          description: The export document, as a download
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
            Content-Disposition: { schema: { type: string, example: 'attachment; filename="tunedeck-export-2026-10-05.json"' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/MeAccountExport' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { description: 'ACCOUNT_DELETING, AUTH_FORBIDDEN or DEVICE_REVOKED', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'API_RATE_LIMITED, 120 reads a minute per account; details.retryAfterSeconds', headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { $ref: '#/components/responses/Error' }
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT, description: 'OIDC access token (iss/aud/exp/RS256 verified via JWKS)' }
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 'Optional on signed-in mutations. The same account, method, path and key runs at most once for 24 hours; a repeat with the same body and If-Match replays the first answer with Idempotent-Replayed: true. Different body → 409 IDEMPOTENCY_KEY_REUSED; still running → 409 IDEMPOTENCY_IN_PROGRESS with Retry-After. Ignored where the answer carries a secret (export links, support codes).'
      schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' }
  schemas:
...
    Error:
      type: object
      required: [code, messageKey, requestId, details]
      properties:
        code:
          type: string
          enum: [AUTH_REQUIRED, AUTH_FORBIDDEN, VALIDATION_FAILED, NOT_FOUND, EXPORT_NOT_READY, IDEMPOTENCY_KEY_REQUIRED, PRECONDITION_REQUIRED, REVISION_MISMATCH, PAYLOAD_TOO_LARGE, DEPENDENCY_UNAVAILABLE, REAUTH_REQUIRED, DEVICE_REVOKED, DEVICE_LIMIT, ROLE_REQUIRED, PUBLISH_BLOCKED, CHECK_TOO_SOON, CHECKER_UNAVAILABLE, RIGHTS_ALREADY_REVOKED, API_RATE_LIMITED, ACCOUNT_DELETING, JOB_RETRY_TOO_SOON, JOB_BUSY, JOB_RETRY_LIMIT, JOB_SUPERSEDED, SYNC_RESET_REQUIRED, PURCHASE_INVALID, PURCHASE_CONFLICT, MFA_REQUIRED, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, SUPPORT_ACCESS_REQUIRED, INTERNAL, DIRECTORY_BLOCK_EXISTS]
        messageKey: { type: string, example: errors.settings.revisionMismatch }
        requestId: { type: string }
        details:
          type: object
          description: 'No tokens or PII. VALIDATION_FAILED → {field?, reason, allowed?}; REVISION_MISMATCH → {currentRevision}; REAUTH_REQUIRED → {maxAgeSeconds}; MFA_REQUIRED → {maxAgeSeconds, scope?}; DEVICE_LIMIT → {maxActiveDevices}; PUBLISH_BLOCKED → {reasons}; IDEMPOTENCY_IN_PROGRESS / CHECK_TOO_SOON / JOB_RETRY_TOO_SOON → {retryAfterSeconds}.'
    Settings:
      type: object
      additionalProperties: false
      required: [theme, language, cellularPolicy]
      properties:
        theme: { type: string, enum: [system, light, dark] }
        language: { type: string, enum: [th, en] }
        cellularPolicy: { type: string, enum: [allow, wifi_only] }
...
    SettingsView:
      type: object
      required: [revision, schemaVersion, settings, updatedAt]
      properties:
        revision: { type: integer, minimum: 0, description: 0 until the first save }
        schemaVersion: { type: integer, const: 1 }
        settings: { $ref: '#/components/schemas/Settings' }
        updatedAt: { type: [string, 'null'], format: date-time, description: null until the first save }
...
    Device:
      type: object
      required: [id, platform, osMajor, appBuild, appliedSettingsRevision, lastSeenAt, lastSyncedAt, revokedAt, overrides, preferencesRevision]
      properties:
        id: { type: string, format: uuid, description: Lowercase }
        platform: { type: string, enum: [ios, android] }
        osMajor: { type: integer }
        appBuild: { type: string }
        appliedSettingsRevision: { type: integer, minimum: 0 }
        lastSeenAt: { type: string, format: date-time, description: Server time of the last check-in; not a location or listening signal }
        lastSyncedAt: { type: [string, 'null'], format: date-time, description: When the device last finished a sync push or pull; null if never }
        revokedAt: { type: [string, 'null'], format: date-time }
        overrides: { $ref: '#/components/schemas/MeSettingsOverrides' }
        preferencesRevision: { type: integer, minimum: 0, description: The device's preferences revision, 0 until first set }
...
    MeAccount:
      type: object
      required: [userId, email, status, createdAt]
      properties:
        userId: { type: string, format: uuid, description: The account code customers give support }
        email: { type: [string, 'null'], maxLength: 320, description: From the access token's email claim at sign-in; null if the token never carried one }
        status: { type: string, enum: [active, deleting, deleted, disabled], description: Always active in practice; other states are refused at sign-in (403) }
        createdAt: { type: string, format: date-time }
...
    MeSettingsOverrides:
      type: object
      additionalProperties: false
      description: Only the keys a device sets differently from the account; {} follows the account.
      properties:
        theme: { type: string, enum: [system, light, dark] }
        language: { type: string, enum: [th, en] }
        cellularPolicy: { type: string, enum: [allow, wifi_only] }
...
    MeDeletionTicket:
      type: object
      required: [ticket, status]
      properties:
        ticket: { type: string, pattern: '^[A-Za-z0-9_-]{43}$', description: 'Secret; poll GET /v1/account-deletions/{ticket}' }
        status: { type: string, const: deleting }
    MeDeletionStatus:
      type: object
      required: [status, requestedAt, deadline, completedAt]
      properties:
        status: { type: string, enum: [deleting, completed, failed], description: 'failed means retrying or dead-lettered (an operator retries it); the account stays locked meanwhile' }
        requestedAt: { type: string, format: date-time }
        deadline: { type: string, format: date-time, description: requestedAt plus 30 days }
        completedAt: { type: [string, 'null'], format: date-time }
...
    MeAccountExport:
      type: object
      required: [format, version, exportedAt, account, settings, devices, favorites, purchases, diagnostics, diagnosticsTruncated, staffRoles]
      description: Everything the service holds about the account (Doc 17 export).
      properties:
        format: { type: string, const: tunedeck-account-export }
        version: { type: integer, const: 1 }
        exportedAt: { type: string, format: date-time }
        account:
          type: object
          required: [id, email, createdAt, locale]
          properties:
            id: { type: string, format: uuid }
            email: { type: [string, 'null'] }
            createdAt: { type: string, format: date-time }
            locale: { type: [string, 'null'] }
        settings: { oneOf: [{ $ref: '#/components/schemas/SettingsView' }, { type: 'null' }], description: null when never saved }
        devices: { type: array, items: { $ref: '#/components/schemas/Device' }, description: Every device, revoked included }
        favorites:
          type: array
          items:
            type: object
            required: [stationId, order, updatedAt]
            properties:
              stationId: { type: string, format: uuid }
              order: { type: integer }
              updatedAt: { type: string, format: date-time }
        purchases:
          type: array
          items:
            type: object
            required: [store, productId, state, environment, purchasedAt, verifiedAt, revokedAt]
            properties:
              store: { type: string, enum: [apple, google] }
              productId: { type: string }
              state: { type: string, enum: [verified, pending, revoked] }
              environment: { type: string, enum: [production, sandbox] }
              purchasedAt: { type: [string, 'null'], format: date-time }
              verifiedAt: { type: [string, 'null'], format: date-time }
              revokedAt: { type: [string, 'null'], format: date-time }
        diagnostics:
          type: array
          items:
            type: object
            required: [id, deviceId, receivedAt, eventCount, events]
            properties:
              id: { type: string, format: uuid }
              deviceId: { type: string, format: uuid }
              receivedAt: { type: string, format: date-time }
              eventCount: { type: integer }
              events: { type: array, items: { $ref: '#/components/schemas/MeDiagnosticEvent' } }
        diagnosticsTruncated: { type: boolean, description: True when more than 10,000 events exist and the rest were left out }
        staffRoles:
          type: array
          items:
            type: object
            required: [role, grantedAt, revokedAt]
            properties:
              role: { type: string, enum: [support, catalog_editor, operator, admin, auditor] }
              grantedAt: { type: string, format: date-time }
              revokedAt: { type: [string, 'null'], format: date-time }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }